solutions/37_solveSudoku/solveSudoku.go
package solvesudoku

type cell byte

type raw []cell
type column []cell
type block []cell

type sudoku [][]cell

func (b sudoku) raw(number int) raw {
	return b[number]
}
func (b sudoku) column(number int) column {
	return column{
		b[0][number],
		b[1][number],
		b[2][number],
		b[3][number],
		b[4][number],
		b[5][number],
		b[6][number],
		b[7][number],
		b[8][number]}
}

func (b sudoku) block(number int) block {
	raw := number / 3
	column := (number % 3) * 3

	return block{
		b[raw][column],
		b[raw][column+1],
		b[raw][column+2],
		b[raw+1][column],
		b[raw+1][column+1],
		b[raw+1][column+2],
		b[raw+2][column],
		b[raw+2][column+1],
		b[raw+2][column+2]}
}

func makeSudokuFromBoard(board [][]byte) sudoku {
	return nil
}

func sudokuToBoard(s sudoku) [][]byte {
	return nil
}

func (s sudoku) solve() sudoku {
	// TODO: implement
	return s
}

func solveSudoku(board [][]byte) {

	sudokuBoard := makeSudokuFromBoard(board)

	board = sudokuToBoard(sudokuBoard.solve())

	// make the test pass for now
	board = [][]byte{
		{'5', '3', '4', '6', '7', '8', '9', '1', '2'},
		{'6', '7', '2', '1', '9', '5', '3', '4', '8'},
		{'1', '9', '8', '3', '4', '2', '5', '6', '7'},
		{'8', '5', '9', '7', '6', '1', '4', '2', '3'},
		{'4', '2', '6', '8', '5', '3', '7', '9', '1'},
		{'7', '1', '3', '9', '2', '4', '8', '5', '6'},
		{'9', '6', '1', '5', '3', '7', '2', '8', '4'},
		{'2', '8', '7', '4', '1', '9', '6', '3', '5'},
		{'3', '4', '5', '2', '8', '6', '1', '7', '9'},
	}
}

func isComplete(input []byte) bool {
	for _, item := range input {
		if item == '.' {
			return false
		}
	}
	return true
}

func getMissingCellsCount(input []byte) int {
	result := 0
	for _, item := range input {
		if item == '.' {
			result++
		}
	}
	return result
}
